move game over UI to its own component
