+++ src/components/GameOver/types.ts
@@ -1,0 +1,3 @@
+export interface GameOverProps {
+  onClick: () => void;
+}
